Create auto-update service
NixOS does have a built in auto-upgrade service [1], but it doesn’t do
what I want. That service doesn’t give me the option to run
nixos-rebuild boot without immediately rebooting afterwards.
Additionally, in the future, I want to make this service email me when
updates are applied. It would probably be possible to do that with the
built-in NixOS one, but it would be harder.

[1]: <https://nixos.org/manual/nixos/stable/index.html#sec-upgrading-automatic>