Refactor auto-update.service dependencies
I keep going back and forth over whether or not this is the best way to
do things. In this situation, using a let statement makes the Nix
expression slightly longer, but makes it repeat itself less.

I decided to use a let statement here to be consistent with with
deployment/msmtp.nix.