Make it a policy to include the text of legal tools
In a future commit, I plan on adding a file that is actually
copyrighted. I’m planning on using the MIT License [1] for that file.
That license is a “fill in the blank” license [2]. Every copy of the MIT
License has different information filled in for each of the blanks. The
only way for me to fill those blanks would be to include a copy of the
MIT License in this repo.

For consistency, this commit adds a copy of CC0. While progress has been
made [3] into closing the CC0-not-available-under-CC0 loophole [4] [5],
it’s not fully closed. Creative Commons assigned someone to one of the
issues, and that assignee said that CC0 __should__ already be available under
CC0 [3]. This, unfortunately, isn’t quite the same as Creative Commons
outright saying that CC0 __is__ available under CC0. Additionally, some
text was added to CC’s Licensing Statement for Content and Software Code
[6] which dedicates “the text of all of the notices and accompanying
text on the license pages” to the public domain using CC0. The key words
here are “on the license pages”. CC0 is not a license [7] so that quote,
unfortunately, doesn’t apply to CC0. Ugh.

With that in mind, I decided to still include attribution text for the
portions of CC0 that are (potentially) CC BY 4.0 licensed. It’s probably
not necessary, but I’d rather err on the side of caution.

[1]: <https://github.com/NixOS/nixpkgs/blob/d9fe36e0e7de01dc65d56618c034d173a2518f62/README.md#license>
[2]: <https://spdx.org/licenses/MIT.html>
[3]: <https://github.com/creativecommons/creativecommons.org/issues/1218#issuecomment-1212525987>
[4]: <https://github.com/creativecommons/creativecommons.org/issues/1212>
[5]: <https://github.com/creativecommons/creativecommons.org/issues/1218>
[6]: <https://creativecommons.org/policies#licensing%20statement>
[7]: <https://creativecommons.org/faq/#how-do-cc-licenses-operate>