Allow tmpfses to be huge
Like I mentioned in the previous commit message, I was trying to build
something and I was running out of space in a tmpfs. This change will
make that less likely.

This change also increases the risk that a program would use a ton of
virtual memory and lock up the system. I estimate that the likelihood of
that happening is pretty low.

This commit also creates a script that can be used to test the new size
limits.