Allow all graphical systems to manage VMs
I want to run a VM on Jason-Laptop-Linux in order to test out ECWolf [1]
bugs. I know that these bugs are present in the nixpkgs version of
ECWolf, but are they in the official ECWolf packages [2]? I’ll need an
Ubuntu [3] VM to find out.

I could run a VM on Jason-Desktop-Linux to test this theory, but, at the
moment, I’m using Jason-Desktop-Linux to try and find a workaround for a
bug that affects Mu [4]. Every time I come up with a potential work-
around, I have to wait a long time (about 350 seconds) to see whether or
not the workaround actually worked. As a result, I don’t want to run a
VM on Jason-Desktop-Linux at the moment (sometimes, I end up having to
reboot the host in order to fix the VM, and I don’t want to interrupt
the testing that I’m doing with Mu).

[1]: <https://maniacsvault.net/ecwolf/>
[2]: <https://maniacsvault.net/ecwolf/download.php>
[3]: <https://ubuntu.com>
[4]: <https://codewith.mu>