Install Transmission on all graphical systems
I’m currently working on testing a commit that will declaratively
configure ECWolf. I’m concerned that the configuration in that commit
won’t work on a fresh install, so I want to test it out in a VM. I’m
using Jason-Desktop-Linux to reproduce a bug with Mu [1] that takes a
very long time to occur, so I don’t want to use Jason-Desktop-Linux to
host a VM.

As a result of that situation, I want to run a VM on Jason-Laptop-Linux.
I have another commit that allows me to run VMs on my laptop, but, when
I went to test that commit, I realized that I needed to download an ISO
to boot the VM. I was going to use this as an opportunity to try to
reproduce a bug with ECWolf [2]. I’m able to reproduce the ECWolf bug on
NixOS, but I’m not sure if the bug is in ECWolf itself or if the bug is
only present in the nixpkgs [3] ecwolf package [4]. In order to
determine which project introduced the bug, I want to set up a VM
running Ubuntu [5] so that I can test out one of the official ECWolf
packages [6].

OK, so I want to download an Ubuntu ISO on Jason-Laptop-Linux.
Technically, I could just use HTTP(S) [7] to download the Ubuntu ISO,
but I would much rather use BitTorent [8]. BitTorent is a more efficient
protocol and using BitTorent allows me to give back to the Ubuntu
community by seeding.

That’s why I want to run Transmission on all graphical systems (both
Jason-Desktop-Linux and Jason-Laptop-Linux). I’m already using
Transmission on Jason-Desktop-Linux. I want to use Transmission on
Jason-Laptop-Linux to download an Ubuntu ISO so that I can create a VM
so that I can try to reproduce an ECWolf bug.

[1]: <https://codewith.mu>
[2]: <https://maniacsvault.net/ecwolf/>
[3]: <https://github.com/NixOS/nixpkgs>
[4]: <https://search.nixos.org/packages?channel=22.05&show=ecwolf&from=0&size=50&sort=relevance&type=packages&query=ecwolf>
[5]: <https://ubuntu.com/>
[6]: <https://maniacsvault.net/ecwolf/download.php>
[7]: <https://httpwg.org/specs/>
[8]: <https://www.bittorrent.org>