Only import Home Manager in machine specific configs
Before this change, .nix files that customized Home Manager specific
properties would import Home Manager. I thought that doing so made sense
because those files relied on Home Manager being imported.

In the future, I want to make it so that some systems import the 22.05
version of Home Manager and that other systems import the 22.11 version
of Home Manager. I want to do that so that I can upgrade each system to
22.11 individually (as opposed to having to upgrade each system all at
once). In other words, I want to implement
<mid:20221025165942.ztowsv7io4kzlwvw@Jason-Laptop-Linux> in the future.

The point of this change is to make it easier to implement per-system
versions of Home Manager. Now that only the machine specific configs
import Home Manager, there’s less places that have to be updated when I
want to move the first system to version 22.11 of NixOS and Home
Manager.