Allow machines to sync with Graphical-Test-VM
9bcbeda (Add support for creating a graphical test VM, 2022-12-18) added
devices to Graphical-Test-VM’s list of Syncthing devices but didn’t add
Graphical-Test-VM to other machine’s list of Syncthing devices. That was
because, at the time, Graphical-Test-VM’s Syncthing ID hadn’t been
generated yet. You have to run Syncthing for the first time to get it to
generate an ID.

Now that Graphical-Test-VM is up and running, it’s time to make other
devices sync with it.

Implements <mid:20221220004614.sb74dry3bbhl3fd7@Jason-Desktop-Linux>.