Update Graphical-Test-VM’s Syncthing ID
The last commit, 32d23cb (Upgrade Graphical-Test-VM to NixOS 22.11,
2022-12-29), changed what version of NixOS Graphical-Test-VM is supposed
to use. I wanted to be sure that the config still worked with NixOS
22.11, so I tested that commit by

1. upgrading Graphical-Test-VM from 22.05 to 22.11 and
2. doing a fresh install of NixOS 22.11 on Graphical-Test-VM.

When you reinstall a system, a new Syncthing ID is generated. This
commit updates Graphical-Test-VM’s Syncthing ID to match the one that
was generated with that new installation.