Update Jason-Lemur-Pro-VM-Test’s Syncthing device ID
I’m currently working on implementing [1] which is a new feature for
Jason’s RHEL Config [2]. While testing part of the implementation of
that feature, I deleted Jason-Lemur-Pro-VM-Test’s Syncthing config and
regenerated it. Regenerating a Syncthing config causes a new device ID
to be generated, so this repo had to be updated in order to keep my
NixOS machines syncing with Jason-Lemur-Pro-VM-Test.

[1]: <mid:20230119212224.rrkcaqdepzeohj5j@Jason-Laptop-Linux>
[2]: <https://jasonyundt.website/gitweb?p=jasons-rhel-config;a=summary>