Allow the creation of macOS VMs
I recently submitted a pull request that updates nixpgs’s ecwolf package
to 1.4.0 [1]. That PR was accepted, but some changes were requested. One
of the changes was to add a comment that explains why the build fails on
macOS [2].

Here’s the problem: I don’t know why the build fails on macOS. I’m going
to set up a VM, install macOS and then try the build to see what
happens. I’m following the instructions on the NixOS Wiki for how to do
this [3], and those instructions say that I have to add stuff to my
NixOS config before I get started.

[1]: <https://github.com/NixOS/nixpkgs/pull/211602>
[2]: <https://github.com/NixOS/nixpkgs/pull/211602#discussion_r1085436433>
[3]: <https://nixos.wiki/wiki/OSX-KVM>