Auto-check test mailserver’s IPv4 connection
For a while, mailserver.test.jasonyundt.email had an IPv6 address, but
couldn’t connect to anything using IPv6. If that ever happens again, I
want to know about it.

This change doesn’t implement a test to check that, but it does create a
framework for adding more tests in the future and making sure that I
know when they fail.

Partially implements
<mid:20230212195123.lt3nzh5yy2gb7ybn@Jason-Desktop-Linux>.