Rename deployment/ to imports/
For one thing, the name “imports” is more clear (it’s where the files
that get imported by configuration.nix are stored). More importantly,
the name deployment causes a tab-completion collision with deploy.sh. As
a result, you would have to type ./d<tab>.<tab> in order to get
“./deploy.sh”. Now, you just have to type ./d<tab>. The main motivation
behind this change is to save time in the future.

Implements <mid:20230214125111.vhwqprqyj55ou374@Jason-Desktop-Linux>.