Update pre-commit
I’m working on a commit in the mailserver-dns branch that requires a
newer version of pre-commit. See the comments that this commit adds for
details.

This commit moves the pre-commit.override part into an overlay.
Originally, I was planning on leaving the pre-commit.override part in
the home.packages list, but I wasn’t able to get that to work with an
overlay. I chose to use an overlay in the first place since I couldn’t
think of any other way to override the version of python311Packages.cffi
that pre-commit ends up using.