Revert "Update pre-commit"
This reverts commit 655adc59607514f8fa064c16da746419bd09d5bd.

That commit updated pre-commit because I was working on some new
pre-commit hooks in a separate branch. I was running into what I thought
was a pre-commit bug, and I thought that updating pre-commit would
potentially fix the bug (or at least, give me an environment that I can
use to report the bug). As it turns out, the pre-commit behavior that I
was experiencing is actually intentional:
<https://github.com/pre-commit/pre-commit/issues/1926>. Newer versions
of pre-commit don’t make the behavior go away.

As a result, I don’t really benefit from importing the newer version of
pre-commit from a PR. Doing so just adds extra complexity, and I don’t
even know what the difference between the two versions is.