Add test for low disk space
If the server runs out of disk space, then it won’t be able to accept
any new mail which would be a problem. Specifically, this test will fail
if any filesystems are more than 85% full. That number is arbitrary, but
the hope is that filesystems will fill up slowly enough that it won’t be
too late by the time this test fails. I might have to lower that
threshold later, though.

Implements <mid:20230220115812.koad73gwgs7de7io@Jason-Desktop-Linux>.