[WIP] Enable DNSSEC
The main advantage of DNSSEC is that it allows people to be confident in
the authenticity of the records that they receive.

Out of all of the DNS security features that I’m going to implement,
I’ve decided to implement this one first. Later, I’m going to enable DNS
over QUIC and use security.acme to generate a cert for that will be used
with DNSSEC. Specifically, I’m going to use a DNS challenge to generate
the cert. I don’t know if the CA will actually check the DNSSEC records
before generating the cert, but it would make sense if they did (it
would prove that the records are indeed true, even if there was a
man-in-the-middle attack).

TODO:
• Open to do item for automatically testing DNSSEC.
• Add documentation for generating DS records and applying them to the
root zone.
• Test (see <https://www.dnssec-validator.cz/> and
<https://zonemaster.labs.nic.cz/>)