Update glue record test for new name servers
Eventually, I want to enable DNSSEC, but, unfortunately, the
Mail-in-a-Box UI doesn’t allow me to add custom DS records. As a result,
I switched to using Gandi LiveDNS name servers for jasonyundt.email. As
a result, the glue records now need to be on those servers, not on
box.jasonyundt.email.

Fixes <mid:20230316143607.u7qqs5nm4x6djimi@Jason-Desktop-Linux>.