Make mode read-only
In general, I make variables read-only unless they need to be changed.
It’s a habit I picked up from Rust. I already do this in the other
pieces of Bash code that are in this repo, so it makes sense to do it in
deploy.sh as well.

Implements <mid:20230317134835.p3vnxtj4isoke32h@Jason-Desktop-Linux>.