Prevent update.sh from prompting the user
Sometimes, you know ahead of time that rebuilding the NixOS
configuration is going to take a while. It’s kind of annoying to walk
away from the machine because it’s going to take a while and then to
come back to see that it’s waiting for you to answer a question from
flatpak.

Implements <mid:20221223123639.53dhb5xnanwigl5g@Jason-Laptop-Linux>.