Move Nix expression for Knot DNS into its own file
The portion of my NixOS system configuration for Knot DNS is already
very long. In a future commit, I’m going to make it even longer. I’m
going to move to an unstable version of Knot DNS so that I can get
access to these changes [1]. Those changes make it so that I can use
DNS-over-QUIC without having to enable XDP. I tried enabling XDP, but I
couldn’t get it to work. Not having DNS-over-QUIC isn’t an option. I
refuse to not allow people to use encryption if they want it, and
DNS-over-QUIC seems to be the best option for encrypting authoritative
name server traffic.

[1]: <https://gitlab.nic.cz/knot/knot-dns/-/merge_requests/1497/commits>