[WIP] Ignore absolute includes in knot.conf
In a future commit, I’m going to make it so that DNS records can be
dynamically updated [1] if a program has access to a secret key [2] [3].
In order to do so, that future commit will use keymgr [4] to generate a
conf file that will be imported by knot.conf. It’s important that that
new conf file won’t be put in this repo so that the secret contained in
it isn’t leaked.

Unfortunately, “knotc conf-check” fails in the conf file includes
another conf file that can’t be found. I don’t expect pre-commit to ever
be run on a system that actually has that file in place, or any other
includes that have an absolute path. Other includes that have an
absolute path would probably also only be available at runtime on the
system where the config is being deployed.

One might ask, why go through all these hoops to enable dynamic DNS in
the first place? I’m enabling dynamic DNS so that I can use
security.acme in configuration.nix to generate an SSL cert. I’m going to
use the SSL cert for DNS over QUIC. I could have used something other
than a DNS challenge to get the cert, but it seems natural to use a DNS
challenge if the server already has a built-in DNS server.

Additionally, I could configure Knot DNS without any sort of
authentication, just as long as the connection comes from localhost.
While that would be relatively secure, it would allow any unprivileged
user to update DNS. Attackers wouldn’t even need to bother doing
privilege escalation because they could just overwrite the DNS records
to point to their own server.

TODO:
• Proofread (Check links)
• Test.

[1]: <https://www.rfc-editor.org/info/rfc2136>
[2]: <https://www.knot-dns.cz/docs/3.2/html/reference.html#key-section>
[3]: <https://www.knot-dns.cz/docs/3.2/html/reference.html#acl-section>
[4]: <https://www.knot-dns.cz/docs/3.2/html/man_keymgr.html>