Allow update-serial-numbers to run on all zones
This change makes it so that if you run update-serial-numbers.py without
any arguments, then it will be run on every zone listed by “git
ls-files”. A future change will make update-serial-numbers get run even
if a zone file hasn’t been changed. This change is in preparation for
that future change. This change makes it easier to implement that future
change.