Fix SOA RNAME
The DNS SOA RNAME field is supposed to correspond to an email address.
For whatever reason, you can’t use an at sign in an RNAME field.
Instead, the first dot gets replaced with an at sign to form the email
address [1].

Before this change, the RNAME was “jason.jasonyundt.email”. I assumed
that that RNAME would get translated into the email address
“jason@jasonyundt.email”. In reality, the part after the first dot is a
relative domain, not an absolute one. In other words, “jasonyundt.email”
was interpreted as being relative to the $ORIGIN. As a result, the RNAME
field was being transformed into
“jason@jasonyundt.email.test.jasonyundt.email”, which isn’t a real email
address. By adding a final dot to the RNAME field, the domain becomes an
absolute one.

Fixes <mid:20230416110356.gkcchdc6i57o4762@Jason-Desktop-Linux>.

[1]: <https://www.nslookup.io/learning/dns-record-types/soa/#soa-record-format-and-fields>