Use SOA values recommended by Zonemaster
The developers of Knot DNS [1] (the authoritative name server that I
use) also created a tool called Zonemaster that checks whether or not a
DNS server is configured properly [2]. That tool recommended that I
change some of the values in my SOA record. I assume that they know
better than me when it comes to DNS, so I’m trusting their judgement
when it comes to these values.

[1]: <https://www.knot-dns.cz/>
[2]: <https://zonemaster.labs.nic.cz/>