Fix NixOS manual section numbers
That part of the installation instructions were written before stable
NixOS had a graphical installer. When the graphical installer was added,
a new subsection (2.2) about the graphical installer was added to the
manual, and the subsections that came after it had their numbers bumped.
This change updates the section numbers used in this repo’s README so
that they line up with the correct sections in the newer versions of the
stable NixOS manual.

Fixes <mid:20221219233451.t4dh42eyg7r5dn5i@Jason-Desktop-Linux>.