Make a command’s output more clear
If the instructions tell you to try and figure out whether or not Nix is
installed, then the command that they have you run should talk about
whether or not Nix is installed, not whether or not nix-build is
installed. The instructions and the command’s output should match.

Implements <mid:20221227174026.ulzzak6duglw7t5e@Jason-Desktop-Linux>.