Add command for choosing random file or directory
I have a long list of bugs that I’ve been meaning to report, but
haven’t gotten around to yet. Each bug is a file, and the files are
organized into directories. The randfd command makes it easy to choose a
random item from that list.

Implements <mid:20230217190757.iyltgtxhwkgupmlv@Jason-Desktop-Linux>.