Make unstable Nixpkgs importable
A future change is going to make a file other than git-graphical.nix use
something from the unstable version of Nixpkgs. This change allows that
future change to be more efficient. I’ll be able to use unstable Nixpkgs
in the future change without having to rewrite the URL or
“builtins.fetchTarball”.