Add build instructions for NixOS 22.11’s manual
Unfortunately, the NixOS project only seems to distribute builds of
manuals for the current stable version and the current unstable version
of NixOS. As of 2023-06-02T06:30-4:00, NixOS 23.05 is the new stable
version of NixOS [1]. As a result, builds of the NixOS 22.11 manual are
no longer available on nixos.org.

Technically, I could just follow the instructions in the NixOS 23.05
manual to install NixOS 22.11, and it would probably work, but I would
rather use the version of the manual that corresponds with the version
of NixOS that I’m going to be installing. It just seems less risky to do
so.

This commit also removes the instructions for building the NixOS 22.05
manual. NixOS 22.05 hasn’t been supported since 2022-12-31 [2], and ever
since d92f90b (Upgrade Jason-Desktop-Linux to NixOS 22.11, 2022-12-30),
none of my machines have used NixOS 22.05.

[1]: <https://discourse.nixos.org/t/nixos-23-05-released/28649?u=jasonyundt>
[2]: <https://discourse.nixos.org/t/nixos-22-11-released/23637?u=jasonyundt>