Make pre-commit workaround compatible with 23.05
Before this change, a let-expression was used to shadow the pkgs
variable. To make this change, I needed access to the old pkgs variable
(the one that got shadowed), so I gave the variable that was shadowing
pkgs a new name. That’s probably what I should have done in the first
place since shadowing can be misleading.