Get wooting-udev-rules from unstable Nixpkgs
Now that this PR [1] is merged, we can get the updated
wooting-udev-rules package directly from Nixpkgs Unstable (rather than
having to get it from a PR). We were already downloading Nixpkgs
Unstable in git-graphical.nix, so we save time by not having to download
that PR’s version of Nixpkgs.

Additionally, the branch that was being downloaded was deleted. This
config continued to build successfully because there was a copy of the
branch in my Nix store, but it was giving me an error, and it would have
failed if I had to reinstall NixOS.

Fixes <mid:ptoatgtbjhempjlveuh2uytep6koupjlugjqjafikdyjf6yfc4@bsdctcc7ykic>.

[1]: <https://github.com/NixOS/nixpkgs/pull/225055>