Create script that nicely shuts down systems
When it comes to powering off a Linux system, there’s too many different
commands that you could possibly use [1] [2] [3] [4]. This new script
will try to do what ever method makes the most sense for the system that
it’s being run on.

On its own, this script isn’t all that useful. Eventually, it will be
used to allow systems to automatically reboot after deploying this
config [5].

[1]: <man:shutdown(8)>
[2]: <man:poweroff(8)>
[3]: <man:systemctl(8)>
[4]: <https://askubuntu.com/q/1871/565319>
[5]: <mid:6jdn7u327fxruiqfjplio4aaj3zgl46fnbxsddrbrwatgisd2t@24yfqk5n5jrw>