Enable polkit on all systems
Before this change, imports/applications/nicely-stop-session.nix was
kind of awkward to use on headless systems. That script would try to run
“systemctl poweroff” or “systemctl reboot” and then fail with a
permission denied error. A workaround would be to run the entire script
using sudo, but that’s not ideal either (don’t run stuff as root unless
you need to!).

This change should allow systemctl to prompt for a password if it needs
additional privileges, which will make nicely-stop-session more
convenient on headless systems.