Remove 30 second delay from nicely-stop-session
When nicely-stop-session was originally wrote, it was being used like
this:

	{ nohup nicely-stop-session reboot & } && disown && exit

This would cause the current terminal session to stop before the
shutdown was initiated. Originally, I thought that this was a good idea
to make sure that the terminal shuts down completely before the system
starts to shutdown. The point of the 30 second wait was to make sure
that the terminal was done exiting before the shutdown starts.

That being said, I no longer think that it makes sense to run
nicely-stop-session that way. On headless systems, systemd should be
able to stop any TTY or SSH sessions gracefully. On graphical systems,
KDE Plasma should close out of the running terminal gracefully. After
all, both the systemctl and qdbus commands terminate before the shutdown
actually starts.

When I wrote 3c5fcd9 (Create quick-deploy.sh, 2023-09-13), I didn’t
bother to use nohup with nicely-stop-session. The main reason why I
created nicely-stop-session was so that I could later create a script
like quick-deploy.sh. Now that quick-deploy.sh exists, I doubt that I’m
going to run nicely-stop-session directly very often, and when I do, I
doubt that I’m going to bother using nohup.

As a result, delaying for 30 seconds won’t be useful, except potentially
in some very rare circumstances. At this point, it’s just a waste of
time.