Potentially bypass polkit using Sudo
When you run “systemctl poweroff” or “systemctl reboot”, it will use
polkit to prompt you for a password. If you run either of those commands
with Sudo, then it might prompt you for a password or it might not (the
credentials might be cached). Potentially not prompting for a password
is a speed improvement.