Automatically upgrade pre-commit hooks
I never remember to run “pre-commit autoupdate”. This hook will
hopefully make it so that I don’t have to remember.

This change makes pre-commit emit a warning about a mutable reference.
That warning is safe to ignore and will go away once the TODO items that
this commit adds are dealt with.