Use Docker version of shfmt pre-commit hook
In a future commit, I’m going to update the shfmt hook to v3.7.0-2 [1].
That new version makes a change to how the main shfmt pre-commit hook
works, and, unfortunately, that change breaks my config. I can no longer
use a custom `args` array with the main `shfmt` hook, but it still works
for the `shfmt-docker` one [2].

This change is in preparation for that future commit that will update
the shfmt hook.

[1]: <https://github.com/scop/pre-commit-shfmt/releases/tag/v3.7.0-2>
[2]: <https://github.com/scop/pre-commit-shfmt/issues/19>