Tell users to potentially build 23.05’s manual
Before this change, the installation instructions would tell users to
potentially build the manual for NixOS 23.11 because the NixOS project
wasn’t distributing manuals for NixOS 23.11 yet. The NixOS project only
distributes manuals for the latest stable version of NixOS and for the
latest unstable version of NixOS [1]. Now that NixOS 23.11 has been
released [2], NixOS 23.05 is no longer the latest stable version of
NixOS, and, as a result, the NixOS project is no longer distributing
built versions of the NixOS 23.05 manual.

[1]: <https://github.com/NixOS/nixos-homepage/blob/a78bec0c0f20c9d1b7cd07686dacb05beeb6f1a6/flake.nix#L135>
[2]: <https://discourse.nixos.org/t/nixos-23-11-released/36210?u=jasonyundt>