Don’t leave result symlink for deploy.sh
Before this commit, deploy.sh would build a package which would put a
result symlink in the root of the repo. After deploy.sh finished
running, the symlink would still be there, even though it wasn’t needed
anymore.

This could have been accomplished by simply removing the symlink, but I
thought that it was more elegant to use nix-shell directly. That way,
the symlink would never be created in the first place.

This commit moves the bulk of the source code for the deploy script from
deploy.nix to imports/applications/deploy.nix. I could have just put
everything in deploy-shell.nix, but I chose to put the script itself
into imports/applications/ to keep things more organized.

I also changed the file that builds the deploy script to do this:

	{ pkgs ? import <nixpkgs> { } }:

instead of this:

	with import <nixpkgs> { };

That change makes the file that builds the deploy script more consistent
with the other files in imports/applications/.