Make quick-deploy.sh check for sudo
All of the scripts in this repo should either automatically get
dependencies using nix-shell [1] or resholve [2]. Unfortunately, scripts
like this don’t seem to work:

	#!/usr/bin/env nix-shell
	#! nix-shell -i bash -p sudo

	sudo echo hi

When I run a script like that on my system, I get this error:

	sudo: <path>/sudo must be owned by uid 0 and have the setuid bit set

For scripts that depend on sudo, I’ve been using an if statement to make
sure that sudo is present. Unfortunately, I had forgotten to put that if
statement in quick-deploy.sh. This change fixes that issue.

[1]: <https://nixos.org/manual/nix/stable/command-ref/nix-shell>
[2]: <https://github.com/abathur/resholve>