Make quick-deploy.sh prompt for password faster
Before this commit, quick-deploy.sh used this shebang:

	#!/usr/bin/env nix-shell
	#! nix-shell -i bash -p nix

Unfortunately, using nix-shell as an interpreter [1] is slow. Before
this commit, it would take a second or two for nix-shell to start. As a
result, the user would have to wait a second or two before entering
their password. This change will (hopefully) eliminate that short delay.

[1]: <https://nixos.org/manual/nix/stable/command-ref/nix-shell#use-as-a--interpreter>