Add configuration.nix files to repo
This change makes it easier to reinstall NixOS. I won’t have to create
my own configuration.nix file. Instead, I’ll be able to just run the
deploy script and have it create the configuration.nix file for me.

Additionally, this change is in preparation for a future commit that
will use nixos-rebuild --target-host [1]. I need to have the entirety of
the remote host’s configuration on my local system in order for
--target-host to work.

[1]: <https://www.mankier.com/8/nixos-rebuild#--target-host>