Remove Flatpak-related code
Ever since 910a958 (Remove Tor Browser, 2024-01-18), none of the systems
have used Flatpak. Removing this code will (hopefully) make
deploy-jasons-nixos-config faster.

Additionally, a future commit is going to rework
deploy-jasons-nixos-config so that I can use it to deploy a config on a
remote system. Creating the future commit is going to be easier if I
don’t have to worry about Flatpak.