Allow arbitrary arguments to be passed to nixos-rebuild
Before this commit, deploy-jasons-nixos-config only allowed you to run
“nixos-rebuild boot” or “nixos-rebuild switch”. This change makes it so
you could use deploy-jasons-nixos-config to run other nixos-rebuild
commands like “nixos-rebuild build” or
“nixos-rebuild boot --specialisation example”.

The main motivation behind this commit is to make writing two future
commits easier. One of those future commits is going to make it so that
you can use deploy-jasons-nixos-config to deploy a configuration on a
remote system. In order to do so, I need to be able to make
deploy-jasons-nixos-config run nixos-rebuild with the --target-host
option [1]. The second future commit is going to create a systemd unit
that will periodically update channels and then rebuild the config in
the background. In order to do so, I need to be able to make
deploy-jasons-nixos-config run “nixos-rebuild build”.

[1]: <https://www.mankier.com/8/nixos-rebuild#--target-host>