Make running nixos-rebuild as root optional
The motivation behind this commit is to make it easier to write two
future commits. Those two future commits will use
deploy-jasons-nixos-config to run
“nixos-rebuild boot --target-host <remote-host>” and
“nixos-rebuild build”. Those two nixos-rebuild commands don’t need to
be run as root.

This commit makes deploy-jasons-nixos-config check to make sure that
JNC_NIXOS_REBUILD_AS_ROOT is set to a valid value. The idea behind that
check is to prevent future mistakes (like typos) from going unnoticed.