Don’t bother copying configuration into /etc/nixos
The main motivation behind this commit is to make it easier to write two
future commits. Both of those future commits involve building the config
on one machine and then deploying that config on another machine. This
commit makes it easier to do that. This commit makes it so that we can
run deploy-jasons-nixos-config without messing up the files in
/etc/nixos. In fact, it makes it so that the files in /etc/nixos don’t
matter at all.