Remove dependency on /etc/nixos
53f1813 (Don’t bother copying configuration into /etc/nixos, 2024-01-18)
should have made the /etc/nixos/ directory totally unnecessary.
Unfortunately, I forgot to remove one of the references to /etc/nixos/.
This commit fixes that problem.

The main motivation behind this change is to make a future commit work.
That future commit will add a remote-deployment script. That script was
failing because it was trying to run nixos-rebuild as a normal user.
Before 53f1813 (Don’t bother copying configuration into /etc/nixos,
2024-01-18), files were copied into /etc/nixos/imports/ and given very
restricted permissions. Specifically, only root was able to read files
inside /etc/nixos/imports/. That remote-deployment script from the
future commit was failing because the script didn’t have access to files
in /etc/nixos/imports/, and it needed access to those files in order to
build the configuration.