Use nix-shell in a less weird way
Before this change, shell scripts in this repo would use the nix-shell
command [1] in a really weird way. Instead of using nix-shell as an
interpreter [2], the scripts in this repo would do this:

	#!/usr/bin/env bash

	nix-shell whatever-shell.nix

whatever-shell.nix would contain the actual code for the script, shoved
into a shellHook attribute [3]. Normally, running
“nix-shell whatever-shell.nix” would give the user a shell prompt. I
didn’t want it to do that, so I put “exit” at the end of each of those
shellHook attributes.

All of that is unintuitive and unnecessarily complicated. The only
reason why I did it that way is because I didn’t realize that I could
use nix-shell as an interpreter while still using a separate
whatever-shell.nix file.

[1]: <https://nixos.org/manual/nix/stable/command-ref/nix-shell>
[2]: <https://nixos.org/manual/nix/stable/command-ref/nix-shell#use-as-a--interpreter>
[3]: <https://nixos.org/manual/nixpkgs/stable/#sec-pkgs-mkShell-attributes>
[4]: 2a8ca2c (Make quick-deploy.sh prompt for password faster, 2024-01-18)