Deploy full config when installing
The idea here is to make setting up a system faster. Before this commit,
I had to run nixos-rebuild immediately after installing a system. This
commit makes it so that I don’t have to do that.

Additionally, this commit was made to make it easier for me to create a
future commit. That future commit will automate the setup process. It’s
going to be easier to do that if the full config gets deployed.