Include prebuilt config on install ISOs
The idea here is to make installing NixOS faster. If the installation
image contains all of the packages needed to build the config, then I
won’t have to wait for things to be downloaded. Additionally, this
commit allows me to build everything on a faster machine, and then
install on a slower machine.