Make IUJ generate a new hardware configuration
Before this commit, you would have to regenerate the hardware
configuration before running install-using-jnc. This commit makes it so
that install-using-jnc automatically regenerates the hardware
configuration for you. This will make installing NixOS faster.

The main motivation behind this commit is to make it easier to write a
future commit. That commit will automate the installation process.