Use Disko on jasonyundt.website.home.arpa
The goal of this commit is to make it easier to reinstall
jasonyundt.website.home.arpa. The easier it is to reinstall that VM, the
easier it will be for me to test out changes that require
reinstallation. Specifically, I’m working on another branch that will
switch systems from ext4 [1] to bcachefs [2]. In order to test out the
changes in that branch, I need to reinstall NixOS in that VM, and I’m
getting sick and tired of following the installation instructions over
and over again. The goal of this commit is to make it bearable to
reinstall NixOS just for the sake of testing out changes.

[1]: <https://www.kernel.org/doc/Documentation/filesystems/ext4.txt>
[2]: <https://bcachefs.org>