Reorganize packages into pkgCollections
Before this commit, there was a .nix file that provided a copy of
Nixpkgs Unstable. That file was located in src/modules despite that fact
that it wasn’t a module. This commit fixes that problem.

Additionally, this commit creates a good place to put any future package
collections. I’m working on commits in other branches that would benefit
from additional package collections.