Compress install ISOs way less
Compression is a waste of time when it comes to installation ISOs
because the data gets used locally. The ISO files will either be read by
a local virtual machine or copied to a USB drive. If I was distributing
ISOs over the Internet, then I would want to use compression in order to
save bandwidth, but I’m not doing that.

I want to just turn off compression entirely, but unfortunately the pull
request that allows you to turn off compression hasn’t made it into
NixOS 23.11 yet [1]. Since I can’t turn off compression entirely, my
next best bet is to choose the algorithm that will compress and
decompress my data as quickly as possible. I haven’t bothered
benchmarking all of the available compression algorithms, but I did find
this page [2] that suggests that lz4 is the fastest. That page is the
only reason why I chose lz4 over any of the other options.

[1]: <https://github.com/NixOS/nixpkgs/pull/287514>
[2]: <https://catchchallenger.herman-brule.com/wiki/Quick_Benchmark:_Gzip_vs_Bzip2_vs_LZMA_vs_XZ_vs_LZ4_vs_LZO#Compression_time>