Set passwords before installing
This will (hopefully) make installing NixOS faster.

Before this commit, deploy.sh would end up running a command like this:

	VAR=whatever command

This commit makes deploy.sh use the env command to do that:

	env VAR=whatever command

I had to do that in order to allow the command to get packed into an
array. Apparently, you Bash doesn’t allow you to do this:

	array=( 'VAR=whatever' 'command' )
	"${array[@]}"

But it does allow you to do this:

	array=( 'env' 'VAR=whatever' 'command' )
	"${array[@]}"

¯\_(ツ)_/¯