Resholve remote-deploy.sh
This ensures that remote-deploy.sh doesn’t have any undeclared
dependencies. Additionally, this commit makes remote-deploy.sh more
consistent with some of the other Bash scripts in this repo
(build-iso.sh, deploy.sh and generate-hardware-configuration.sh). That
consistency is going to make it easier for me to create a future commit.