Create “Unattended Install” boot menu entry
Hopefully, this will make reinstalling NixOS faster. I could have just
made the main installation options do an unattended install, but I
want to also be able to use installation ISOs to rescue systems.

The way that this commit adds a boot menu entry is convoluted, but I
couldn’t think of a better way of doing it. See this NixOS Discourse
post [1] for details.

[1]: <https://discourse.nixos.org/t/how-do-i-add-boot-menu-entries-to-an-install-iso/39748?u=jasonyundt>