Fix resholve not resolving commands run by sudo
For whatever reason, when resholve encounters something like this:

	sudo echo hi

it doesn’t resolve the echo command. I’m not sure why it doesn’t resolve
that command, but I have a feeling that it’s related to this issue [1].

This commit works around that resholve limitation by pre-resolving
commands that get executed by sudo. The main motivation behind this
commit is to make creating a future commit easier. In order for that
future commit to work, this command:

	sudo nixos-install

must be transformed into this command:

	sudo <path-to-nixos-install>

because nixos-install won’t be on the PATH.

[1]: <https://github.com/abathur/resholve/issues/113>